+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_prims.py
@@ -35,8 +35,9 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     ]
     # TODO StageManager: We have LSS plugin names in the flux ext because of this system
     compatible_widgets: list[str] = [
-        "PrimTreeWidgetPlugin",
+        "CustomTagsWidgetPlugin",
         "FocusInViewportActionWidgetPlugin",
-        "IsVisibleActionWidgetPlugin",
         "IsCaptureStateWidgetPlugin",
+        "IsVisibleActionWidgetPlugin",
+        "PrimTreeWidgetPlugin",
     ]