+++ source/extensions/omni.flux.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.7.0]
+### Added
+- Added `custom_tags_list` widget
+
 ## [1.6.1]
 ### Fixed
 - Fixed tests flakiness