+++ source/extensions/omni.flux.stage_manager.window/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.0]
+### Changed
+- Updated the test schema to includes new plugins
+
 ## [1.2.2]
 ### Fixed
 - Fixed tests flakiness